+++ src/atom/atomizer.ts
@@ -1,9 +1,9 @@
-import Mustache from "mustache";
+import Atom from "./atom";
 import FileManager from "../utils/filemanager";
 import * as htmlparser2 from "htmlparser2"
 import Logger from "../utils/logger";
 import fs from "fs";
-import { Component, String } from "../components";
+import { Component, String, Style } from "../components";
 
 /**
  * An Atomizer template can be a string or null.
@@ -28,27 +28,66 @@ export default class Atomizer {
      * Only the templated from this folder will be preloaded loaded.
      * If there are many template folders then those components will have to be loaded manually.
      * or change the template folder and preload the templates again.
+     * 
+     * @author Ansh Sharma
      */
-    public static templateFolder: string = "./src/templates/";
+    public static templateFolders: [{
+        baseDir: string,
+        htmlDir?: string,
+        cssDir?: string,
+        jsDir?: string
+    }] = [{ baseDir: "./src/templates/", htmlDir: "", cssDir: "css", jsDir: "js" }];
+
+    public static templateFilesToInclude: string[] = [];
 
     /**
      * The preloaded templates (the default ones)
      */
-    public static templates:{[key:string]:AtomizerTemplate} = Atomizer.preloadTemplates();
+    public static templates: { [key: string]: AtomizerTemplate }[] = Atomizer.preloadTemplates();
+
 
     /**
      * Load a template from the template folder
      * @param filename The name of the template file
      * @returns New atomizer template
      */
-    public static loadTemplate(filename: string): AtomizerTemplate {
-        const fm = new FileManager(Atomizer.templateFolder);
+    public static loadTemplate(filename: string, templateFolderIndex: number): AtomizerTemplate {
+        const fm = new FileManager(Atomizer.templateFolders[templateFolderIndex].baseDir);
         const newName = filename;
         const template = fm.readFile(newName);
         if (template == null) {
             Logger.error(__filename, `Template ${newName} not found`);
             return null;
         }
+
+        let baseName = newName.split(".")[0];
+        if (fs.existsSync(Atomizer.templateFolders[templateFolderIndex].baseDir
+            + Atomizer.templateFolders[templateFolderIndex].cssDir
+            + "/"
+            + baseName
+            + ".css")) {
+            Atomizer.templateFilesToInclude.push(
+                Atomizer.templateFolders[templateFolderIndex].baseDir
+                + Atomizer.templateFolders[templateFolderIndex].cssDir
+                + "/"
+                + baseName
+                + ".css");
+        }
+
+        if (fs.existsSync(Atomizer.templateFolders[templateFolderIndex].baseDir
+            + Atomizer.templateFolders[templateFolderIndex].jsDir
+            + "/"
+            + baseName
+            + ".js")) {
+            Atomizer.templateFilesToInclude.push(
+                Atomizer.templateFolders[templateFolderIndex].baseDir
+                + Atomizer.templateFolders[templateFolderIndex].jsDir
+                + "/"
+                + baseName
+                + ".js");
+        }
+
+
         Logger.info(__filename, `Template ${newName} loaded`);
         return template;
     }
@@ -57,16 +96,93 @@ export default class Atomizer {
      * Preload All the templates from the template folder.
      * @returns A dictionary of all the preloaded templates with their names as keys.
      */
-    public static preloadTemplates():{[key:string]:AtomizerTemplate}{
-        const templates:{[key:string]:AtomizerTemplate} = {};
-        const files = fs.readdirSync(Atomizer.templateFolder);
-        files.forEach(file => {
-            let newKey = file.split(".")[0];
-            const template = Atomizer.loadTemplate(file);
-            if(template != null)
-                templates[newKey] = template;
+    public static preloadTemplates(): { [key: string]: AtomizerTemplate }[] {
+        const fm = new FileManager();
+        const templates: { [key: string]: AtomizerTemplate } = {};
+        return Atomizer.templateFolders.map((templateFolder, index) => {
+            const files = fs.readdirSync(templateFolder.baseDir, { withFileTypes: true });
+            files.forEach(file => {
+                if (!file.isDirectory()) {
+                    let newKey = file.name.split(".")[0];
+                    const template = Atomizer.loadTemplate(file.name, index);
+                    if (template != null)
+                        templates[newKey] = template;
+                }
+                else if (file.isDirectory()) {
+                    console.log(file.name);
+                    if (file.name === templateFolder.htmlDir) {
+                        const htmlFiles = fs.readdirSync(templateFolder.baseDir + templateFolder.htmlDir, { withFileTypes: true });
+
+                        htmlFiles.forEach(htmlFile => {
+                            if (!htmlFile.isDirectory()) {
+                                let newKey = htmlFile.name.split(".")[0];
+                                const template = Atomizer.loadTemplate(templateFolder.htmlDir + "/" + htmlFile.name, index);
+                                if (template != null)
+                                    templates[newKey] = template;
+                            }
+                        });
+                    }
+                }
+            });
+
+            Logger.info(__filename, `Templates loaded from ${templateFolder.baseDir}`);
+            return templates;
+
         });
-        return templates;
+    }
+
+    /**
+     * Build the component tree from the atom.
+     * @param atom The atom to be parsed and built into a component tree
+     * @returns The root component of the component tree.
+     */
+    public static buildComponentTreeFromAtom(atom: Atom): Component {
+        return Atomizer.buildComponentTree(atom.toString());
+    }
+
+    public static addTemplateFolder(templateFolder: {
+        baseDir: string,
+        htmlDir?: string,
+        cssDir?: string,
+        jsDir?: string
+    }) {
+        if (!Atomizer.templateFolders.find((folder) => folder.baseDir === templateFolder.baseDir)) {
+            Atomizer.templateFolders.push(templateFolder);
+            console.log(Atomizer.templateFolders)
+            Atomizer.templates.push(Atomizer.preloadTemplates()[Atomizer.templateFolders.length - 1]);
+        }
+        else {
+            Logger.warn(__filename, `Template folder ${templateFolder.baseDir} already exists`);
+        }
+    }
+
+    public static getTemplate(templateName: string, templateFolderIndex?: 0): AtomizerTemplate {
+        const template = Atomizer.templates[templateFolderIndex || 0][templateName];
+        if (!template) {
+            const newTemplate = Atomizer.templates.map((templateFolder) => {
+                if (templateFolder[templateName]) {
+                    return templateFolder[templateName];
+                }
+            }).join("");
+            if (!newTemplate) {
+                const new_Template = Atomizer.loadTemplate(templateName, templateFolderIndex || 0);
+                if (new_Template) {
+                    Atomizer.templates[templateFolderIndex || 0][templateName] = new_Template;
+                    return new_Template;
+                }
+                else {
+                    Logger.error(__filename, `Template ${templateName} not found`);
+                    return "";
+                }
+            }
+            else {
+                return newTemplate;
+            }
+        }
+        else {
+            return template;
+        }
+
     }
 
     /**