+++ src/components/scripts.ts
@@ -1,3 +1,5 @@
+import Logger from "../utils/logger";
+import ScriptManager from "../utils/scriptmanager";
 
 
 /**
@@ -45,6 +47,10 @@ export default class Script {
         this.url = url;
         this.type = type;
         this.params = params;
+        if(!ScriptManager.isScriptAdded(this))
+            ScriptManager.addscript(this);
+        else
+            Logger.warn(__filename, `Script ${this.toString()} already added`);
     }
 
     /**