+++ src/templates/js/header.js
@@ -1,0 +1,6 @@
+const button = document.querySelector("#header-collapse");
+const menu = document.querySelector("header ul");
+
+button.addEventListener("click", function () {
+  menu.classList.toggle("hidden");
+});