+++ src/utils/filemanager.ts
@@ -8,7 +8,7 @@ import Logger from './logger';
  * A class dedicated for handling files and directories
  * This class is responsible for managing project folders, copying files from source to build folders.
  * @author Sujal Choudhari <sujalchoudhari@gmail.com>
- */
+*/
 export default class FileManager {
     /**
      * Absolute path to current working directory
@@ -22,8 +22,8 @@ export default class FileManager {
      */
     constructor(basepath: string = "") {
         if (basepath == "")
-            this.basePath = path.resolve(process.cwd());
-        else this.basePath = path.resolve(basepath);        
+            this.basePath = path.resolve(__dirname, '../../');
+        else this.basePath = path.resolve(basepath);
     }
 
     /**
@@ -32,7 +32,7 @@ export default class FileManager {
      * @returns The abosolute path of the file or directory.
      */
     public getAbsolutePath(relativePath: string): string {
-        return path.join(this.basePath, relativePath);
+        return path.resolve(this.basePath, relativePath);
     }
 
     /**
@@ -146,12 +146,14 @@ export default class FileManager {
      * Copy a file to a destination
      * @param srcPath Source file path
      * @param destPath Destination file path
+     * @param srcPathType The type of the source path, relative or absolute. Default is relative
      */
-    public copyFile(srcPath: string, destPath: string): void {
+    public copyFile(srcPath: string, destPath: string, srcPathType : "relative" | "absolute" = "relative"): void {
         let contents = this.readFile(srcPath);
-
-        if (contents !== null)
-            this.createFile(destPath, contents);
+        let fileName = srcPath.split("/").pop();
+        if (contents !== null) {
+            this.createFile(destPath + "/"+ fileName , contents);
+        }
         else {
             Logger.warn(__filename, "Failed to read file");
             return;
@@ -164,9 +166,13 @@ export default class FileManager {
      * @param contents The contents of the file to create with.
      */
     public createFile(filePath: string, contents: string): void {
-        const absolutePath = this.getAbsolutePath(filePath);
-        console.log(absolutePath);
-        fs.writeFileSync(absolutePath, contents);
+        try {
+            fs.writeFileSync(filePath, contents);
+        }
+        catch (err : any) {
+            const absolutePath = this.getAbsolutePath(filePath);
+            fs.writeFileSync(absolutePath, contents);
+        }
     }
 
     /**