+++ CHANGELOG.md
@@ -5,9 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.0.1-alpha.4] 
+## [Work in Progress]
+
+### Added
+- `Atomizer.buildComponentTreeFromAtom` added.
+
 
 # [1.0.1-alpha.3] - 30-03-2023
-
+## [Unreleased]
+The project is going under various changes. 
+The code is not stable and may break at any time.
 ### Added
 - Atom Class.
 - Atomizer Class.